Allow narrative API to return unmodified markdown file
This is in preparation for client-side parsing of narrative files. A optional `type` URL query allows the same API to return the markdown file (see the documentation changes in this commit for more details).

There are other possibilities, such as using the content-type field in the header, but I felt this was the clearest.