[client-narratives] Allow client to gracefully fall-back to JSON
Here we add the framework to (eventually) parse the response from `getNarrative` as a markdown file and interpret this client-side.

To facilitate backward compatibility in the cases where a server does not respect the `?type=md` query and continues to return a JSON, if the file doesn't look like markdown we will attempt to parse it as JSON.

To mimic such a server, we explicitly request "type=json" in this commit, as the code to actually do the client-side md-parsing is not yet implemented.